move parameter file
