+++ parameter.go
@@ -1,0 +1,49 @@
+package h2sc
+
+type PARAMETER []byte
+
+var (
+	PARAMETER_OUTPUT_MUTE_UNMUTE = PARAMETER("00000")
+	PARAMETER_OUTPUT_MUTE_ALL    = PARAMETER("+0001")
+	PARAMETER_OUTPUT_MUTE_VIDEO  = PARAMETER("+0002")
+	PARAMETER_OUTPUT_MUTE_AUDIO  = PARAMETER("+0003")
+
+	PARAMETER_OUTPUT_FREEZE_ENABLE  = PARAMETER("+0001")
+	PARAMETER_OUTPUT_FREEZE_DISABLE = PARAMETER("00000")
+
+	PARAMETER_OUTPUT_ASPECT_ASPECTKEEP    = PARAMETER("00000")
+	PARAMETER_OUTPUT_ASPECT_FULL          = PARAMETER("+0001")
+	PARAMETER_OUTPUT_ASPECT_HORIZONAL_FIT = PARAMETER("+0002")
+	PARAMETER_OUTPUT_ASPECT_VERTICAL_FIT  = PARAMETER("+0003")
+
+	PARAMETER_OUTPUT_FORMAT_1080i5994 = PARAMETER("00000")
+	PARAMETER_OUTPUT_FORMAT_480i      = PARAMETER("+0001")
+	PARAMETER_OUTPUT_FORMAT_720p5994  = PARAMETER("+0004")
+	PARAMETER_OUTPUT_FORMAT_1080p5994 = PARAMETER("+0005")
+	PARAMETER_OUTPUT_FORMAT_1080p60   = PARAMETER("+0006")
+	PARAMETER_OUTPUT_FORMAT_1080p30   = PARAMETER("+0008")
+	// PARAMETER_OUTPUT_FORMAT_1080i5994 = PARAMETER("+0016") // Same as 000000
+	// PARAMETER_OUTPUT_FORMAT_480i = PARAMETER("+0017") // same as +0001
+
+	PARAMETER_OUTPUT_SEAMLESS_BEHAVIOR_FREEZE_SEAMLESS    = PARAMETER("00000")
+	PARAMETER_OUTPUT_SEAMLESS_BEHAVIOR_BLACK_CONNECT      = PARAMETER("+0001")
+	PARAMETER_OUTPUT_SEAMLESS_BEHAVIOR_BLACK_FADE_CONNECT = PARAMETER("+0002")
+	PARAMETER_OUTPUT_SEAMLESS_BEHAVIOR_FRAMELOCK          = PARAMETER("0001")
+
+	PARAMETER_OUTPUT_ROTATE_NORMAL                                     = PARAMETER("00000")
+	PARAMETER_OUTPUT_ROTATE_90DEGREES_COUNTER_CLOCKWISE                = PARAMETER("+0001")
+	PARAMETER_OUTPUT_ROTATE_90DEGREES_CLOCKWISE                        = PARAMETER("+0002")
+	PARAMETER_OUTPUT_ROTATE_180DEGREES                                 = PARAMETER("+0003")
+	PARAMETER_OUTPUT_MIRROR_HORIZON                                    = PARAMETER("+0004")
+	PARAMETER_OUTPUT_MIRROR_VERTICAL                                   = PARAMETER("+0005")
+	PARAMETER_OUTPUT_ROTATE_90DEGREES_COUNTER_CLOCKWISE_MIRROR_VERICAL = PARAMETER("+0006") // fuck.
+	PARAMETER_OUTPUT_ROTATE_90DEGREES_CLOCKWISE_MIRROR_VERICAL         = PARAMETER("+0007")
+
+	PARAMETER_OUTPUT_POWERSAVE_1MIN  = PARAMETER("00000") // 00000 or +0004
+	PARAMETER_OUTPUT_POWERSAVE_5MIN  = PARAMETER("+0001") // +0001 or +0005
+	PARAMETER_OUTPUT_POWERSAVE_10MIN = PARAMETER("+0002") // +0002 or +0006
+	PARAMETER_OUTPUT_POWERSAVE_NEVER = PARAMETER("+0003")
+
+	PARAMETER_OUTPUT_TESTPATTERN_DISABLE = PARAMETER("00000")
+	PARAMETER_OUTPUT_TESTPATTERN_ENABLE  = PARAMETER("+0001")
+)
